+++ rusl/src/unistd/open.rs
@@ -1,6 +1,6 @@
 use sc::syscall;
 
-use crate::compat::unix_str::AsUnixStr;
+use crate::string::unix_str::AsUnixStr;
 use crate::platform::{Fd, AT_FDCWD};
 
 /// Openflags accepted by the [open syscall](https://man7.org/linux/man-pages/man2/open.2.html)
@@ -32,34 +32,34 @@ transparent_bitflags! {
 
 /// Mode accepted by the [open syscall](https://man7.org/linux/man-pages/man2/open.2.html)
 transparent_bitflags! {
-    pub struct Mode: i32 {
-        const S_IRWXU = linux_rust_bindings::S_IRWXU; // 00700 user read write exec
-        const S_IRUSR = linux_rust_bindings::S_IRUSR; // 00400 user Read
-        const S_IWUSR = linux_rust_bindings::S_IWUSR; // 00200 user write
-        const S_IXUSR = linux_rust_bindings::S_IXUSR;  // 00100 user execute
-        const S_IRWXG = linux_rust_bindings::S_IRWXG;  // 00070 group read write exec
-        const S_IRGRP = linux_rust_bindings::S_IRGRP;  // 00040 group read
-        const S_IWGRP = linux_rust_bindings::S_IWGRP;  // 00020 group write
-        const S_IXGRP = linux_rust_bindings::S_IXGRP;   // 00010 group exec
-        const S_IRWXO = linux_rust_bindings::S_IRWXO;   // 00007 other read write exec
-        const S_IROTH = linux_rust_bindings::S_IROTH;   // 00004 other read
-        const S_IWOTH = linux_rust_bindings::S_IWOTH;   // 00002 other write
-        const S_IXOTH = linux_rust_bindings::S_IXOTH;   // 00001 other execute
+    pub struct Mode: u32 {
+        const S_IRWXU = linux_rust_bindings::S_IRWXU as u32; // 00700 user read write exec
+        const S_IRUSR = linux_rust_bindings::S_IRUSR as u32; // 00400 user Read
+        const S_IWUSR = linux_rust_bindings::S_IWUSR as u32; // 00200 user write
+        const S_IXUSR = linux_rust_bindings::S_IXUSR as u32;  // 00100 user execute
+        const S_IRWXG = linux_rust_bindings::S_IRWXG as u32;  // 00070 group read write exec
+        const S_IRGRP = linux_rust_bindings::S_IRGRP as u32;  // 00040 group read
+        const S_IWGRP = linux_rust_bindings::S_IWGRP as u32;  // 00020 group write
+        const S_IXGRP = linux_rust_bindings::S_IXGRP as u32;   // 00010 group exec
+        const S_IRWXO = linux_rust_bindings::S_IRWXO as u32;   // 00007 other read write exec
+        const S_IROTH = linux_rust_bindings::S_IROTH as u32;   // 00004 other read
+        const S_IWOTH = linux_rust_bindings::S_IWOTH as u32;   // 00002 other write
+        const S_IXOTH = linux_rust_bindings::S_IXOTH as u32;   // 00001 other execute
 
         // Linux specific bits
-        const S_ISUID = linux_rust_bindings::S_ISUID; // 0004000 set-user-ID bit
-        const S_ISGID = linux_rust_bindings::S_ISGID; // 0002000 set-group-ID bit
-        const S_ISVTX = linux_rust_bindings::S_ISVTX; // 0001000 set-sticky bit
+        const S_ISUID = linux_rust_bindings::S_ISUID as u32; // 0004000 set-user-ID bit
+        const S_ISGID = linux_rust_bindings::S_ISGID as u32; // 0002000 set-group-ID bit
+        const S_ISVTX = linux_rust_bindings::S_ISVTX as u32; // 0001000 set-sticky bit
 
         // File specific bits
-        const S_IFIFO  = linux_rust_bindings::S_IFIFO;
-        const S_IFCHR  = linux_rust_bindings::S_IFCHR;
-        const S_IFDIR  = linux_rust_bindings::S_IFDIR;
-        const S_IFBLK  = linux_rust_bindings::S_IFBLK;
-        const S_IFREG  = linux_rust_bindings::S_IFREG;
-        const S_IFLNK  = linux_rust_bindings::S_IFLNK;
-        const S_IFSOCK = linux_rust_bindings::S_IFSOCK;
-        const S_IFMT   = linux_rust_bindings::S_IFMT;
+        const S_IFIFO  = linux_rust_bindings::S_IFIFO as u32;
+        const S_IFCHR  = linux_rust_bindings::S_IFCHR as u32;
+        const S_IFDIR  = linux_rust_bindings::S_IFDIR as u32;
+        const S_IFBLK  = linux_rust_bindings::S_IFBLK as u32;
+        const S_IFREG  = linux_rust_bindings::S_IFREG as u32;
+        const S_IFLNK  = linux_rust_bindings::S_IFLNK as u32;
+        const S_IFSOCK = linux_rust_bindings::S_IFSOCK as u32;
+        const S_IFMT   = linux_rust_bindings::S_IFMT as u32;
     }
 }
 
@@ -69,9 +69,9 @@ transparent_bitflags! {
 /// See above, errors are converted into an Err with the corresponding error code
 pub fn open(path: impl AsUnixStr, flags: OpenFlags) -> crate::Result<Fd> {
     path.exec_with_self_as_ptr(|ptr| {
-        let res = unsafe { syscall!(OPENAT, AT_FDCWD, ptr, flags.bits()) } as i32;
+        let res = unsafe { syscall!(OPENAT, AT_FDCWD, ptr, flags.bits()) };
         bail_on_below_zero!(res, "`OPENAT` syscall failed");
-        Ok(res)
+        Ok(res as i32)
     })
 }
 
@@ -82,9 +82,9 @@ pub fn open(path: impl AsUnixStr, flags: OpenFlags) -> crate::Result<Fd> {
 #[inline]
 pub fn open_mode(path: impl AsUnixStr, flags: OpenFlags, mode: Mode) -> crate::Result<Fd> {
     path.exec_with_self_as_ptr(|ptr| {
-        let res = unsafe { syscall!(OPENAT, AT_FDCWD, ptr, flags.bits(), mode.bits()) } as i32;
+        let res = unsafe { syscall!(OPENAT, AT_FDCWD, ptr, flags.bits(), mode.bits()) };
         bail_on_below_zero!(res, "`OPENAT` syscall failed");
-        Ok(res)
+        Ok(res as i32)
     })
 }
 
@@ -94,9 +94,9 @@ pub fn open_mode(path: impl AsUnixStr, flags: OpenFlags, mode: Mode) -> crate::R
 /// see above
 pub fn open_at(dir: Fd, path: impl AsUnixStr, flags: OpenFlags) -> crate::Result<Fd> {
     path.exec_with_self_as_ptr(|ptr| {
-        let res = unsafe { syscall!(OPENAT, dir, ptr, flags.bits()) } as i32;
+        let res = unsafe { syscall!(OPENAT, dir, ptr, flags.bits()) };
         bail_on_below_zero!(res, "`OPENAT` syscall failed");
-        Ok(res)
+        Ok(res as i32)
     })
 }
 
@@ -111,9 +111,9 @@ pub fn open_at_mode(
     mode: Mode,
 ) -> crate::Result<Fd> {
     path.exec_with_self_as_ptr(|ptr| {
-        let res = unsafe { syscall!(OPENAT, dir, ptr, flags.bits(), mode.bits()) } as i32;
+        let res = unsafe { syscall!(OPENAT, dir, ptr, flags.bits(), mode.bits()) };
         bail_on_below_zero!(res, "`OPENAT` syscall failed");
-        Ok(res)
+        Ok(res as i32)
     })
 }
